src/inbox_container/requirements.txt
azure-functions>=1.22.1,<2.0.0
azure-storage-blob>=12.25.1
azure-storage-queue>=12.12.0
azure-identity>=1.21.0
azure-monitor-opentelemetry>=1.6.6
pdf2image>=1.17.0
python-dotenv>=1.1.0
langchain>=0.3.23
pydantic-core>=2.33.1
pydantic>=2.11.2
rich>=14.0.0
par-ai-core>=0.1.24
